Fixing mainline-crashing typo and streamlining quartz selection
I accidentally had if pm.has_pf instead of if pm.fix_quartz_banding;
so, if you didn't fix banding but had pyfits, you'd crash the mainline.
Also, there's now an option to pre-select the quartz frames for each
exposure. It's not locked in, but it makes quartz selection much quicker.