Fixing bug that occurred if same standard observed twice in one night
If a standard was observed twice (low and high airmass, e.g.), the same value
would be written to pm.calib_stars. Then, in standard_match, there would be
two identical entries, when there should be two different entries. Now, it
looks to see if there is another star of that name already in pm.calib_list
before inserting a new one; if so, it iterates through possible new names.