Background masks can now be set once for all images of an object
The background step now has a unique flag, default=True. For standard stars, which
will shift positions, this is good; it means that you will define the mask for each
position. For galaxies that don't move, however, it means that you're doing the same
mask three or more times in a row. By setting unique=False, you define a mask
once and it fits the backgrounds from there. This has been implemented into the
pipeline.